Former-commit-id: f5c491364b39570a3897d297b76bbf6cd3d5e660

+++ htdocs/bimpcommercial/core/triggers/interface_99_modBimpcommercial_PostProcess.class.php
@@ -21,19 +21,23 @@ class InterfacePostProcess extends BimpCommTriggers
             if (empty($obj_errors)) {
                 switch ($action_name) {
                     case 'CREATE':
-                        $errors = $bimpObject->onCreate($warnings);
+                        if(method_exists($bimpObject, 'onCreate'))
+                            $errors = $bimpObject->onCreate($warnings);
                         break;
 
                     case 'VALIDATE':
-                        $errors = $bimpObject->onValidate($warnings);
+                        if(method_exists($bimpObject, 'onValidate'))
+                            $errors = $bimpObject->onValidate($warnings);
                         break;
 
                     case 'UNVALIDATE':
-                        $errors = $bimpObject->onUnvalidate($warnings);
+                        if(method_exists($bimpObject, 'onUnvalidate'))
+                            $errors = $bimpObject->onUnvalidate($warnings);
                         break;
 
                     case 'DELETE':
-                        $errors = $bimpObject->onDelete($warnings);
+                        if(method_exists($bimpObject, 'onDelete'))
+                            $errors = $bimpObject->onDelete($warnings);
                         break;
                 }
